organizing code correctly, and wiring up blog sitemap
